Codespace Cat
Create the basic design for the portfolio website of CodeSpace Cat. Links to the About Me, Skills, Catfolio, and Contact pages have been added to the header and navigation. created a profile picture and an introductory essay for the "About Me" section. Web development talents and other entertaining cat-related abilities are listed in the newly added talents section. created a table in the Catfolio area to display projects, each with a thumbnail, name, date, and description. included a footer with copyright information and a Contact form for visitors to use to get in contact.